Some workarounds that is supposed to fix isues regarding to when available options change versus when the drop down menu is shown/hidden. Now no sudden updates while the list is being shown/hidden should occour.
This has been fixed by dropping the fade, when the drop menu is being hidden, only when it is being shown - and also by having the ability to force the menu to be hidden (which is useful to use right before doing an operation that is known to cause it to be hidden).

Also there's some minor code cleanup, mainly an unused variable being removed.